optimized the operation of the system

+++ test.py
@@ -6,36 +6,42 @@ import time, os, zipfile, shutil
 
 def conv(upload):
     archive = f'media/{upload}'
-    with zipfile.ZipFile(archive, 'r') as zip_file:
-        zip_file.extractall(archive.split('.')[0])
-
-
     file_name = archive.split('.')[0]
-    only_ima = [f for f in listdir(file_name) if isfile(join(file_name, f)) and '.IMA' in join(file_name, f)]
+    ff = file_name.split('/')[-1]
+    last_files = [f for f in listdir('media') if ff in f and '++' in f and '.zip' not in f]
+    if len(last_files) == 0:
+        with zipfile.ZipFile(archive, 'r') as zip_file:
+            zip_file.extractall(archive.split('.')[0])
 
-    directory = f'{file_name}++{int(time.time()) + 86400}'
-    os.makedirs(directory, exist_ok=True)
+        only_ima = [f for f in listdir(file_name) if isfile(join(file_name, f)) and '.IMA' in join(file_name, f)]
 
-    # number = 1
-    # last_description = ''
-    count = 0
+        directory = f'{file_name}++{int(time.time()) + 86400}'
+        os.makedirs(directory, exist_ok=True)
 
-    for name in only_ima:
-        ima_converter = IMAConverter(join(file_name, name))
-        metadata = ima_converter.IMA_MetaData_ToDict()
+        # number = 1
+        # last_description = ''
+        count = 0
 
-        # if last_description != metadata.SeriesDescription:
-        #     number = 1
+        for name in only_ima:
+            ima_converter = IMAConverter(join(file_name, name))
+            metadata = ima_converter.IMA_MetaData_ToDict()
 
-        path_to_image = f"{directory}/{str(metadata.SeriesDescription).split('] ')[-1].replace(' ', '_')} {count}.png"
-        # number += 1
-        count += 1
+            # if last_description != metadata.SeriesDescription:
+            #     number = 1
 
-        # last_description = metadata.SeriesDescription
+            path_to_image = f"{directory}/{str(metadata.SeriesDescription).split('] ')[-1].replace(' ', '_')} {count}.png"
+            # number += 1
+            count += 1
 
-        ima_converter.save_IMA_ImageFrame(path_to_image)
+            # last_description = metadata.SeriesDescription
 
-    shutil.rmtree(file_name)
+            ima_converter.save_IMA_ImageFrame(path_to_image)
+
+        shutil.rmtree(file_name)
+
+    else:
+        os.rename(f'media/{last_files[-1]}', f'{file_name}++{int(time.time()) + 86400}')
+        directory = f'{file_name}++{int(time.time()) + 86400}'
 
     old_files = [f for f in listdir('media') if '++' in f]
